% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{basic_package_sticker}
\alias{basic_package_sticker}
\title{Creating and initializing a basic package sticker}
\usage{
basic_package_sticker(package_name)
}
\arguments{
\item{package_name}{A \code{character}, the package name.}
}
\value{
No return value, but runs \code{\link[usethis]{use_logo}} in the end.
}
\description{
This function sets up a basic package sticker.
}
