% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_list_of_lists.R
\name{check_list_of_lists}
\alias{check_list_of_lists}
\alias{assert_list_of_lists}
\alias{test_list_of_lists}
\title{Check list of lists}
\usage{
check_list_of_lists(x, len = NULL)

assert_list_of_lists(
  x,
  len = NULL,
  .var.name = checkmate::vname(x),
  add = NULL
)

test_list_of_lists(x, len = NULL)
}
\arguments{
\item{x}{[\code{any}]\cr
Object to check.}

\item{len}{[\code{integer(1)}]\cr
Exact expected length of \code{x}.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
Same as documented in \code{\link[checkmate]{check_list}}.
}
\description{
These functions check whether the input is a list that contains list
elements.
}
\examples{
L <- list(list(1), list(2), 3)
check_list_of_lists(L)
test_list_of_lists(L)
\dontrun{
assert_list_of_lists(L)
}
}
\seealso{
Other list helpers:
\code{\link[=merge_lists]{merge_lists()}}
}
\concept{list helpers}
\keyword{validation}
