% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddirichlet_cpp}
\alias{ddirichlet_cpp}
\title{Compute density of Dirichlet distribution}
\usage{
ddirichlet_cpp(x, concentration, log = FALSE)
}
\arguments{
\item{x}{A \code{numeric}, a weight vector of length \code{p}.
Each vector element must be between \code{0} and \code{1}.
The sum of the vector elements must be \code{1}.}

\item{concentration}{A \code{numeric}, the concentration vector of length \code{p}.}

\item{log}{A \code{logical}, if \code{TRUE} the logarithm of the density value is
returned.
By default, \code{log = FALSE}.}
}
\value{
A \code{numeric}, the density value.
}
\description{
This function computes the density of a Dirichlet distribution.
}
\details{
This function performs no input checks. See \code{\link{ddirichlet}}
for the version with input checks.
}
\keyword{internal}
