% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helpers.R
\name{find_closest_year}
\alias{find_closest_year}
\title{Find the closest year to a given date}
\usage{
find_closest_year(date)
}
\arguments{
\item{date}{A \code{date} in the format of \code{"YYYY-MM-DD"}.}
}
\value{
An \code{integer}, the closest year to the input date.
}
\description{
This function takes a date as input and returns the closest year.
}
\examples{
find_closest_year(as.Date("2022-07-15"))
find_closest_year(as.Date("2022-01-01"))
}
