% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_markov.R
\name{hmm}
\alias{hmm}
\alias{print.hmm}
\title{Define HMM model}
\usage{
hmm(states, sdd)

\method{print}{hmm}(x, ...)
}
\arguments{
\item{states}{An \code{integer}, the number of states.}

\item{sdd}{A \code{character}, defining the type of state-dependent distribution.
Print \code{sdds_dictionary} for possible values.}

\item{x}{An object of class \code{hmm}.}

\item{...}{Currently not used.}
}
\value{
An object of class \code{hmm}.
}
\description{
This function defines a hidden Markov model.
}
