% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lists.R
\name{merge_lists}
\alias{merge_lists}
\title{Merge named lists}
\usage{
merge_lists(...)
}
\arguments{
\item{...}{One or more named \code{list}(s).}
}
\value{
A \code{list}.
}
\description{
This function merges \code{list}s based on their element names. Elements are
only included in the final output \code{list}, if no former \code{list} has
contributed an element with the same name.
}
\examples{
merge_lists(list("a" = 1, "b" = 2), list("b" = 3, "c" = 4, "d" = NULL))
}
\seealso{
Other list helpers:
\code{\link[=check_list_of_lists]{check_list_of_lists()}}
}
\concept{list helpers}
\keyword{transformation}
