% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeli-package.R
\docType{package}
\encoding{UTF-8}
\name{oeli-package}
\alias{oeli}
\alias{oeli-package}
\title{oeli: Some Utilities for Developing Data Science Software}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of general-purpose helper functions that I (and maybe others) find useful when developing data science software. Includes tools for simulation, data transformation, input validation, and more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/oeli}
  \item \url{http://loelschlaeger.de/oeli/}
  \item Report bugs at \url{https://github.com/loelschlaeger/oeli/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com}

}
\keyword{internal}
