% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Build permutations}
\usage{
permutations(x)
}
\arguments{
\item{x}{[\code{atomic()}]\cr
Any \code{vector}.}
}
\value{
A \code{list} of all permutations of \code{x}.
}
\description{
This function creates all permutations of a given \code{vector}.
}
\examples{
permutations(1:3)
permutations(LETTERS[1:3])
}
\references{
Modified version of \url{https://stackoverflow.com/a/20199902/15157768}.
}
\seealso{
Other vector helpers:
\code{\link[=check_numeric_vector]{check_numeric_vector()}},
\code{\link[=check_probability_vector]{check_probability_vector()}},
\code{\link[=chunk_vector]{chunk_vector()}},
\code{\link[=equidistant_vectors]{equidistant_vectors()}},
\code{\link[=insert_vector_entry]{insert_vector_entry()}},
\code{\link[=map_indices]{map_indices()}},
\code{\link[=match_numerics]{match_numerics()}},
\code{\link[=split_vector_at]{split_vector_at()}},
\code{\link[=subsets]{subsets()}},
\code{\link[=vector_occurrence]{vector_occurrence()}}
}
\concept{vector helpers}
\keyword{indexing}
