% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_things.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Draw from Dirichlet distribution}
\usage{
rdirichlet(n = 1, concentration)
}
\arguments{
\item{n}{An \code{integer}, the number of samples.}

\item{concentration}{The non-negative concentration vector of length \code{p}.}
}
\value{
If \code{n = 1} a \code{vector} of length \code{p}, else
a \code{matrix} of dimension \code{n} times \code{p} with samples as rows.
}
\description{
This function draws from a Dirichlet distribution.
}
\examples{
rdirichlet(concentration = 1:3)
rdirichlet(n = 4, concentration = 1:2)

}
