% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdirichlet_cpp}
\alias{rdirichlet_cpp}
\title{Draw from Dirichlet distribution}
\usage{
rdirichlet_cpp(concentration)
}
\arguments{
\item{concentration}{The non-negative concentration vector of length \code{p}.}
}
\value{
A column vector of length \code{p}, the random draw.
}
\description{
This function draws from a Dirichlet distribution.
}
\details{
This function performs no input checks. See \code{\link{rdirichlet}}
for the version with input checks.
}
\keyword{internal}
