% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{renv_development_packages}
\alias{renv_development_packages}
\title{Using development packages when working with \code{{renv}}}
\usage{
renv_development_packages(
  packages = c("covr", "devtools", "DT", "markdown", "R.utils", "yaml"),
  file_name = "development_packages"
)
}
\arguments{
\item{packages}{A \code{character} \code{vector} of package names.}

\item{file_name}{A single \code{character}, the name for the \code{.R} file.}
}
\value{
No return value, but it writes a file to the project root and adds an
entry to the \code{.Rbuildignore} file.
}
\description{
This function creates a file that loads development packages so that
\code{{renv}} can detect and write them to the lockfile.
}
