% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_markov.R
\name{sdd_call}
\alias{sdd_call}
\title{Call a state-dependent distribution}
\usage{
sdd_call(sdd, par_sdd, mode, ...)
}
\arguments{
\item{sdd}{An element of \code{sdds_dictionary}.}

\item{par_sdd}{A named \code{list} of parameters for the state-dependent distribution.}

\item{mode}{Either \code{"density"}, \code{"distribution"}, \code{"quantile"}, or \code{"rng"}.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric}.
}
\description{
This helper function evaluates the density, distribution, quantile, or random
number generator of a state-dependent distribution.
}
\keyword{internal}
