% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_markov.R
\docType{data}
\name{sdds_dictionary}
\alias{sdds_dictionary}
\title{Dictionary of state-dependent distributions}
\format{
The \code{sdds_dictionary} is an \code{\link{Dictionary}} object.
}
\usage{
sdds_dictionary
}
\description{
The \code{sdds_dictionary} object is a dictionary of currently implemented
state-dependent distributions.
}
\keyword{datasets}
