% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_regressors}
\alias{simulate_regressors}
\title{Simulate regressor values}
\usage{
simulate_regressors(
  labels = c("A", "B", "C"),
  n = 100,
  marginals = list(),
  correlation = diag(length(labels)),
  verbose = FALSE
)
}
\arguments{
\item{labels}{(\code{character()})\cr
Unique labels for the regressors.}

\item{n}{(\code{integer(1)})\cr
The number of values per regressor.}

\item{marginals}{(\code{list()})\cr
Optionally marginal distributions for the regressors. If not specified,
standard normal distributions are used. Each list entry must be named
according to a regressor label, and the following distributions can be
specified:}

\item{correlation}{(\code{matrix()})\cr
A correlation matrix of dimension \code{length(labels)}, where the
\code{(i, j)}-th entry defines the correlation between regressor
\code{labels[i]} and \code{labels[j]}.}

\item{verbose}{(\code{logical(1)})\cr
Whether to print information about the simulated regressors.}
}
\value{
A \code{data.frame} with \code{n} rows and \code{length(labels)} columns.
}
\description{
This function simulates regressor values from various marginal distributions
with custom correlations.
}
\details{
TODO
}
