% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{unknown_error}
\alias{unknown_error}
\title{Handling of an unknown error}
\usage{
unknown_error(issue_link = "https://github.com/loelschlaeger/oeli/issues")
}
\arguments{
\item{issue_link}{A \code{character}, the URL to a GitHub issues site.}
}
\value{
No return value, but it throws an error.
}
\description{
This function reacts to an unknown error by throwing an error and linking
to a GitHub issues site with the request to submit an issue.
}
