% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dirichlet.R
\name{ddirichlet_cpp}
\alias{ddirichlet_cpp}
\alias{rdirichlet_cpp}
\alias{ddirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution}
\usage{
ddirichlet_cpp(x, concentration, log = FALSE)

rdirichlet_cpp(concentration)

ddirichlet(x, concentration, log = FALSE)

rdirichlet(n = 1, concentration)
}
\arguments{
\item{x}{A \code{numeric}, a weight vector of length \code{p}.
Each vector element must be between \code{0} and \code{1}.
The sum of the vector elements must be \code{1}.}

\item{concentration}{A \code{numeric}, the concentration vector of length \code{p}.}

\item{log}{A \code{logical}, if \code{TRUE} the logarithm of the density value is
returned.
By default, \code{log = FALSE}.}

\item{n}{An \code{integer}, the number of samples.}
}
\value{
TODO If \code{n = 1} a \code{vector} of length \code{p}, else
a \code{matrix} of dimension \code{n} times \code{p} with samples as rows.
A \code{numeric}, the density value.
}
\description{
This function computes the density of a Dirichlet distribution.
}
\details{
TODO input checks
}
\examples{
x <- c(0.5, 0.3, 0.2)
concentration <- 1:3
ddirichlet(x = x, concentration = concentration)
ddirichlet(x = x, concentration = concentration, log = TRUE)

rdirichlet(concentration = 1:3)
rdirichlet(n = 4, concentration = 1:2)
}
\concept{distribution helpers}
\keyword{distribution}
