% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{Dictionary}
\alias{Dictionary}
\title{Dictionary R6 Object}
\description{
Provides a simple key-value interface based on R6.
}
\examples{
# TODO
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{keys}}{TODO}

\item{\code{alias}}{TODO}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dictionary-new}{\code{Dictionary$new()}}
\item \href{#method-Dictionary-add}{\code{Dictionary$add()}}
\item \href{#method-Dictionary-get}{\code{Dictionary$get()}}
\item \href{#method-Dictionary-remove}{\code{Dictionary$remove()}}
\item \href{#method-Dictionary-alias_keys}{\code{Dictionary$alias_keys()}}
\item \href{#method-Dictionary-print}{\code{Dictionary$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-new}{}}}
\subsection{Method \code{new()}}{
initializing a \code{Dictionary} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$new(
  key,
  alias,
  values,
  value_types,
  allow_overwrite = FALSE,
  alias_allowed = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{TODO}

\item{\code{key}}{TODO}

\item{\code{alias}}{TODO}

\item{\code{values}}{TODO}

\item{\code{value_types}}{TODO}

\item{\code{allow_overwrite}}{TODO}

\item{\code{alias_allowed}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new \code{Dictionary} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-add"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-add}{}}}
\subsection{Method \code{add()}}{
adding an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$add(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly the \code{Dictionary} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-get"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-get}{}}}
\subsection{Method \code{get()}}{
getting elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$get(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{TODO}

\item{\code{key}}{TODO}

\item{\code{value}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the selected value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-remove"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-remove}{}}}
\subsection{Method \code{remove()}}{
removing elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$remove(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{TODO}

\item{\code{key}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly the \code{Dictionary} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-alias_keys"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-alias_keys}{}}}
\subsection{Method \code{alias_keys()}}{
getting keys by alias
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$alias_keys(alias)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alias}}{TODO}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{character}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dictionary-print"></a>}}
\if{latex}{\out{\hypertarget{method-Dictionary-print}{}}}
\subsection{Method \code{print()}}{
printing details of the dictionary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dictionary$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{currently not used}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly the \code{Dictionary} object
}
}
}
