% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_helpers.R
\name{function_defaults}
\alias{function_defaults}
\title{Get default function arguments}
\usage{
function_defaults(f, exclude = NULL)
}
\arguments{
\item{f}{A \code{function}.}

\item{exclude}{A \code{character} of argument names to exclude. Can be \code{NULL} (default)
to not exclude any argument names.}
}
\value{
A named \code{list}.
}
\description{
This function returns the default function arguments (if any).
}
\examples{
f <- function(a, b = 1, c = "", ...) { }
function_defaults(f)
function_defaults(f, exclude = "b")

}
