% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_diagonal_indices.R
\name{matrix_diagonal_indices}
\alias{matrix_diagonal_indices}
\title{Get indices of matrix diagonal}
\usage{
matrix_diagonal_indices(n)
}
\arguments{
\item{n}{An \code{integer}, the matrix dimension.}
}
\value{
An \code{integer} \code{vector}.
}
\description{
This function returns the indices of the diagonal elements of a quadratic
matrix.
}
\examples{
n <- 4
matrix(1:n^2, n, n)
matrix_diagonal_indices(n)

}
