% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oeli-package.R
\docType{package}
\name{oeli-package}
\alias{oeli}
\alias{oeli-package}
\title{oeli: Lennart Oelschläger's R Code}
\description{
Some general utilities that I and maybe others find useful.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/oeli}
  \item Report bugs at \url{https://github.com/loelschlaeger/oeli/issues}
}

}
\keyword{internal}
