% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_helpers.R
\name{permutations}
\alias{permutations}
\title{Build permutations}
\usage{
permutations(x)
}
\arguments{
\item{x}{Any \code{vector}.}
}
\value{
A \code{list} of all permutations of \code{x}.
}
\description{
This function creates all permutations of a given \code{vector}.
}
\examples{
permutations(1:3)
permutations(LETTERS[1:3])

}
\references{
Modified version of \url{https://stackoverflow.com/a/20199902/15157768}.
}
