% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_development_packages.R
\name{renv_development_packages}
\alias{renv_development_packages}
\title{Using development packages when working with \code{{renv}}}
\usage{
renv_development_packages(
  packages = c("covr", "devtools", "DT", "markdown", "R.utils", "yaml"),
  file_name = "development_packages"
)
}
\arguments{
\item{packages}{[\code{character()}]\cr
Package names.}

\item{file_name}{[\code{character(1)}]\cr
The name for the \code{.R} file.}
}
\value{
No return value, but it writes a file to the project root and adds an
entry to the \code{.Rbuildignore} file.
}
\description{
This function creates a file that loads development packages so that
\code{{renv}} can detect and write them to the lockfile.
}
\seealso{
Other package helpers: 
\code{\link{Dictionary}},
\code{\link{Storage}},
\code{\link{identical_structure}()},
\code{\link{input_check_response}()},
\code{\link{match_arg}()},
\code{\link{package_logo}()},
\code{\link{print_data.frame}()},
\code{\link{print_matrix}()},
\code{\link{system_information}()},
\code{\link{unexpected_error}()},
\code{\link{user_confirm}()}
}
\concept{package helpers}
\keyword{packaging}
