% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_correlation_matrix.R
\name{sample_correlation_matrix}
\alias{sample_correlation_matrix}
\title{Sample correlation matrix}
\usage{
sample_correlation_matrix(dim, df = dim, scale = diag(dim))
}
\arguments{
\item{dim}{An \code{integer}, the dimension.}

\item{df}{An \code{integer} greater or equal \code{dim}, the degrees of freedom of
the inverse Wishart distribution.}

\item{scale}{A covariance \code{matrix} of dimension \code{dim}, the scale matrix of the
inverse Wishart distribution.}
}
\value{
A correlation \code{matrix}.
}
\description{
This function samples a correlation matrix by sampling a covariance matrix
from an inverse Wishart distribution and transforming it to a correlation
matrix.
}
\examples{
sample_correlation_matrix(dim = 3)

}
