% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_things.R
\name{sample_transition_probability_matrix}
\alias{sample_transition_probability_matrix}
\title{Sample transition probability matrices}
\usage{
sample_transition_probability_matrix(dim, state_persistent = TRUE)
}
\arguments{
\item{dim}{An \code{integer}, the matrix dimension.}

\item{state_persistent}{Set to \code{TRUE} (default) to put more probability on the diagonal.}
}
\value{
A transition probability \code{matrix}.
}
\description{
This function returns a random, squared matrix of dimension \code{dim}
that fulfills the properties of a transition probability matrix.
}
\examples{
sample_transition_probability_matrix(dim = 3)

}
