% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{Simulator}
\alias{Simulator}
\title{Simulator R6 Object}
\description{
Creates a simulation setup, where a function \code{f} is evaluated \code{runs} times,
optionally at each combination of input values.

Provides some convenience (see below for more details):
\itemize{
\item Simulation results can be restored from a backup if the R session crashes.
\item Failed simulation cases can be re-run.
\item More simulation runs can be conducted after the initial simulation.
\item Parallel computation and progress updates are supported.
}
}
\details{
\subsection{Getting started}{
\enumerate{
\item Initialize a new simulation setup via \code{object <- Simulator$new()}.
\item Define function \code{f} and (optionally) arguments via
\code{object$define(f = f, ...)}.
\item Call \code{object$go(runs)} for \code{runs} simulation of \code{f} evaluated at each
parameter combination.
\item Access the results via \code{object$results}.
\item The field \code{object$cases} lists all simulation cases, including those
already resolved and those still pending.
}

See the examples section.
}

\subsection{Backup}{

TODO (with example for defining and using backup)
}

\subsection{Re-run}{

TODO
}

\subsection{More runs}{

TODO
}

\subsection{Parallel computation}{

By default, simulations run sequentially. But since they are independent,
they can be parallelized to decrease computation time. To enable parallel
computation, use the \href{https://future.futureverse.org/}{\code{{future}} framework}.
For example, run
\preformatted{
future::plan(future::multisession, workers = 4)
}
in advance for computation in 4 parallel R sessions.
}

\subsection{Progress updates}{

Use the \href{https://progressr.futureverse.org/}{\code{{progressr}} framework} to
get progress updates. For example, run the following in advance:
\preformatted{
progressr::handlers(global = TRUE)
progressr::handlers(
  progressr::handler_progress(format = ">> :percent, :eta to go :message")
)
}
}
}
\examples{
f <- function(x, y = 1) { Sys.sleep(runif(1)); x + y + rnorm(1, sd = 0.1) }
object <- Simulator$new(verbose = TRUE)
object$define(f = f, x = as.list(1:2))
object$go(runs = 2)
object$results
object$cases
}
\seealso{
Other simulation helpers: 
\code{\link{correlated_regressors}()},
\code{\link{ddirichlet_cpp}()},
\code{\link{dmvnorm_cpp}()},
\code{\link{dtnorm_cpp}()},
\code{\link{dwishart_cpp}()},
\code{\link{simulate_markov_chain}()}
}
\concept{simulation helpers}
\keyword{simulation}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{results}}{[\code{tibble}, read-only]\cr
The simulation results.}

\item{\code{cases}}{[\code{tibble}, read-only]\cr
The simulation cases.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulator-new}{\code{Simulator$new()}}
\item \href{#method-Simulator-define}{\code{Simulator$define()}}
\item \href{#method-Simulator-go}{\code{Simulator$go()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-new}{}}}
\subsection{Method \code{new()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$new(
  use_backup = NULL,
  verbose = getOption("verbose", default = FALSE)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_backup}}{[\code{NULL} | \code{character(1)}]\cr
TODO}

\item{\code{verbose}}{[\code{logical(1)}]\cr
TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-define"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-define}{}}}
\subsection{Method \code{define()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$define(f, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{[\code{function}]\cr
TODO}

\item{\code{...}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-go"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-go}{}}}
\subsection{Method \code{go()}}{
TODO
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$go(
  runs = 0,
  backup = FALSE,
  path = paste0("backup_", format(Sys.time(), "\%Y-\%m-\%d-\%H-\%M-\%S"))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runs}}{[\code{integer(1)}]\cr
TODO}

\item{\code{backup}}{[\code{logical(1)}]\cr
TODO}

\item{\code{path}}{[\code{character(1)}]\cr
TODO}
}
\if{html}{\out{</div>}}
}
}
}
