% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_information.R
\name{system_information}
\alias{system_information}
\title{General system level information}
\usage{
system_information()
}
\value{
A \code{list} with elements:
\itemize{
\item \code{maschine}, the model name of the device
\item \code{cores}, the number of cores
\item \code{ram}, the size of the RAM
\item \code{os}, the operating system
\item \code{rversion}, the R version used
}
}
\description{
This function returns a \code{list} of general system level information.
}
\examples{
system_information()
}
\seealso{
Other package helpers: 
\code{\link{Dictionary}},
\code{\link{Storage}},
\code{\link{identical_structure}()},
\code{\link{input_check_response}()},
\code{\link{match_arg}()},
\code{\link{package_logo}()},
\code{\link{print_data.frame}()},
\code{\link{print_matrix}()},
\code{\link{renv_development_packages}()},
\code{\link{unexpected_error}()},
\code{\link{user_confirm}()}
}
\concept{package helpers}
\keyword{validation}
