% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unexpected_error.R
\name{unexpected_error}
\alias{unexpected_error}
\title{Handling of an unexpected error}
\usage{
unexpected_error(
  msg = "Oops, an unexpected error occurred.",
  issue_link = "https://github.com/loelschlaeger/oeli/issues"
)
}
\arguments{
\item{msg}{[\code{character(1)}]\cr
An error message.}

\item{issue_link}{[\code{character(1)}]\cr
The URL to an issues site.}
}
\value{
No return value, but it throws an error.
}
\description{
This function reacts to an unexpected error by throwing an error and linking
to an issue site with the request to submit an issue.
}
\seealso{
Other package helpers:
\code{\link{Dictionary}},
\code{\link{Storage}},
\code{\link[=check_missing]{check_missing()}},
\code{\link[=find_namespace_calls]{find_namespace_calls()}},
\code{\link[=find_pkg_functions]{find_pkg_functions()}},
\code{\link[=identical_structure]{identical_structure()}},
\code{\link[=input_check_response]{input_check_response()}},
\code{\link[=match_arg]{match_arg()}},
\code{\link[=package_logo]{package_logo()}},
\code{\link[=print_data.frame]{print_data.frame()}},
\code{\link[=print_matrix]{print_matrix()}},
\code{\link[=system_information]{system_information()}},
\code{\link[=user_confirm]{user_confirm()}}
}
\concept{package helpers}
\keyword{packaging}
