% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_helpers.R
\name{unexpected_error}
\alias{unexpected_error}
\title{Handling of an unexpected error}
\usage{
unexpected_error(
  msg = "We are sorry, an unexpected error occured.",
  issue_link = "https://github.com/loelschlaeger/oeli/issues"
)
}
\arguments{
\item{msg}{A \code{character}, an error message.}

\item{issue_link}{A \code{character}, the URL to a GitHub issues site.}
}
\value{
No return value, but it throws an error.
}
\description{
This function reacts to an unexpected error by throwing an error and linking
to a GitHub issues site with the request to submit an issue.
}
