% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_confirm.R
\name{user_confirm}
\alias{user_confirm}
\title{User confirmation}
\usage{
user_confirm(question = "Question?", default = FALSE)
}
\arguments{
\item{question}{[\code{character(1)}]\cr
The binary question to ask. It should end with a question mark.}

\item{default}{[\code{logical(1)}]\cr
The default decision.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
This function asks in an interactive question a binary question.
}
\seealso{
Other package helpers:
\code{\link{Dictionary}},
\code{\link{Storage}},
\code{\link[=check_missing]{check_missing()}},
\code{\link[=find_namespace_calls]{find_namespace_calls()}},
\code{\link[=find_pkg_functions]{find_pkg_functions()}},
\code{\link[=identical_structure]{identical_structure()}},
\code{\link[=input_check_response]{input_check_response()}},
\code{\link[=match_arg]{match_arg()}},
\code{\link[=package_logo]{package_logo()}},
\code{\link[=print_data.frame]{print_data.frame()}},
\code{\link[=print_matrix]{print_matrix()}},
\code{\link[=system_information]{system_information()}},
\code{\link[=unexpected_error]{unexpected_error()}}
}
\concept{package helpers}
\keyword{validation}
